+++ packages/bill_sqflite_api/lib/src/bill_sqflite_api.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
+import 'dart:developer';
 
 import 'package:bill/bill.dart';
 import 'package:path/path.dart';
+import 'package:rxdart/rxdart.dart';
 import 'package:sqflite/sqflite.dart';
 
 /// {@template bill_sqflite_api}
@@ -13,6 +15,8 @@ class BillSqfliteApi {
 
   Database? _db;
 
+  final _billsController = BehaviorSubject<List<Bill>>.seeded(const []);
+
   Future<Database> get db async {
     _db ??= await initDb();
     return _db!;
@@ -23,17 +27,33 @@ class BillSqfliteApi {
       final databasesPath = await getDatabasesPath();
       final path = join(databasesPath, 'bills.db');
       final db = await openDatabase(path, version: 1, onCreate: _onCreate);
+      final result = await db.rawQuery(
+        'SELECT * FROM bills ORDER BY id',
+      );
+      final bills = result.map(Bill.fromMap).toList();
+      _billsController.add(bills);
       return db;
     } catch (e) {
-      print('Error initializing database: $e');
+      log('Error initializing database: $e');
       rethrow; // Rethrow the error to see the stack trace
     }
   }
 
-  void _onCreate(Database db, int version) async {
+  Future<void> close() async {
+    final dbClient = await db;
+    try {
+      await dbClient.close();
+    } catch (e) {
+      log('Error closing BillSqfliteApi: $e');
+      rethrow;
+    }
+  }
+
+  Future<void> _onCreate(Database db, int version) async {
     await db.execute('''
       CREATE TABLE bills (
         id INTEGER PRIMARY KEY,
+        bill_name TEXT,
         type TEXT,
         qr TEXT,
         name TEXT,
@@ -46,21 +66,19 @@ class BillSqfliteApi {
         date_created TEXT
       )
     ''');
+    await db.execute('''
+      CREATE TABLE currencies (
+        name TEXT PRIMARY KEY
+      )
+    ''');
+    await db.execute('''
+      CREATE TABLE tags (
+        name TEXT PRIMARY KEY
+      )
+    ''');
   }
 
-  Future<int> insertBill(Bill bill) async {
-    final dbClient = await db;
-    return dbClient.insert('bills', await bill.toMap());
-  }
-
-  Future<List<Bill>> fetchAllBills() async {
-    final dbClient = await db;
-    final result = await dbClient.rawQuery(
-      'SELECT * FROM bills ORDER BY id',
-    );
-    final bills = result.map(Bill.fromMap).toList();
-    return bills;
-  }
+  Stream<List<Bill>> getBills() => _billsController.asBroadcastStream();
 
   Future<List<Map<String, dynamic>>> getBill(int id) async {
     final dbClient = await db;
@@ -72,7 +90,23 @@ class BillSqfliteApi {
     return dbClient.query('bills', columns: ['id']);
   }
 
+  Future<int> insertBill(Bill bill) async {
+    final dbClient = await db;
+    final result = await dbClient.insert('bills', await bill.toMap());
+    final bills = [..._billsController.value, bill];
+    _billsController.add(bills);
+    return result;
+  }
+
   Future<int> updateBill(Bill bill) async {
+    final bills = [..._billsController.value];
+    final billIndex = bills.indexWhere((b) => b.id == bill.id);
+    if (billIndex != -1) {
+      bills[billIndex] = bill;
+    } else {
+      bills.add(bill);
+    }
+    _billsController.add(bills);
     final dbClient = await db;
     return dbClient.update(
       'bills',
@@ -83,17 +117,50 @@ class BillSqfliteApi {
   }
 
   Future<int> deleteBill(int id) async {
+    final bills = [..._billsController.value];
+    final billIndex = bills.indexWhere((b) => b.id == id);
+    if (billIndex == -1) {
+      throw Exception('Bill not found');
+    }
+    bills.removeAt(billIndex);
+    _billsController.add(bills);
     final dbClient = await db;
     return dbClient.delete('bills', where: 'id = ?', whereArgs: [id]);
   }
 
-  Future<void> close() async {
+  Future<List<String>> getCurrencies() async {
     final dbClient = await db;
-    try {
-      await dbClient.close();
-    } catch (e) {
-      print(e);
-      rethrow;
+    final result =
+        await dbClient.rawQuery('SELECT DISTINCT name FROM currencies');
+    if (result.isEmpty) {
+      return [];
     }
+    return result.map((e) => e['name']! as String).toList();
+  }
+
+  Future<void> setCurrencies(List<String> currenciesList) async {
+    final dbClient = await db;
+    final batch = dbClient.batch();
+
+    for (final currency in currenciesList) {
+      batch.insert('currencies', {'name': currency});
+    }
+    await batch.commit();
+  }
+
+  Future<List<String>> getTags() async {
+    final dbClient = await db;
+    final result = await dbClient.rawQuery('SELECT DISTINCT name FROM tags');
+    return result.map((e) => e['name']! as String).toList();
+  }
+
+  Future<void> setTags(List<String> tagsList) async {
+    final dbClient = await db;
+    final batch = dbClient.batch();
+
+    for (final tag in tagsList) {
+      batch.insert('tags', {'name': tag});
+    }
+    await batch.commit();
   }
 }
